add update for posts

+++ src/use-cases/post/delete-post.js
@@ -1,5 +1,8 @@
 export default function makeDeletePost({ postsDb }) {
     return async function deletePost(id) {
+        if (!id) {
+            throw new Error('You must supply an id.')
+        }
         return await postsDb.remove(id);
     }
 }
